+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.3.0] - 2022-01-01
+### Added
+-   Total of likes (incrementable with like toggler function).
+-   Title of the current media inside lightbox.
+
 ## [0.3.0] - 2021-12-29
 ### Added
 -   Sorting system (by number of likes, date or title).